refactor: consolidate imports and improve type safety in usuarios store
Consolidate imports from `../types/supabase` into `../index` to reduce redundancy. Improve type safety in `UsuariosStore` by defining explicit types for the store and its methods. Refactor `mostrarUsuarios` and `editartemamonedauser` methods for better readability and error handling.
